게시글 권한 추가

+++ src/main/webapp/jsp/article/detail.jsp
@@ -6,6 +6,7 @@
 <%
 
 	Map<String, Object> articleMap = (Map<String, Object>) request.getAttribute("articleMap");
+	String loginedMemberLoginId = (String) request.getAttribute("loginedMemberLoginId");
 
 %>    
     
@@ -23,16 +24,26 @@
 		<a href="<%= request.getContextPath() %>/home/main">메인</a>
 	</div>
 	<div>
+
 		<h2><%= articleMap.get("id") %> 번 상세보기</h2>
 		<div><%= articleMap.get("updateDate") %></div>
+		<div><%= articleMap.get("writerName") %></div>
 		<div><%= articleMap.get("title") %></div>
 		<div><%= articleMap.get("body") %> </div>
 
 	</div>
 	<div>
 		<a href="list">목록</a>
-		<a href="modify?id=<%= articleMap.get("id") %>">수정</a>
-		<a href="delete?id=<%= articleMap.get("id") %>" onclick = "if(confirm('정말 삭제하시겠습니까?') == false) return false;" >삭제</a>
+		<%
+		if (loginedMemberLoginId != null && loginedMemberLoginId.equals(articleMap.get("writerName") )) {
+		%>
+			<a href="modify?id=<%= articleMap.get("id") %>">수정</a>
+			<a href="delete?id=<%= articleMap.get("id") %>" onclick = "if(confirm('정말 삭제하시겠습니까?') == false) return false;" >삭제</a>
+		<%
+		}
+		%>
+		
+	
 	</div>
 </body>
 </html>
